feat: make telemetry ativation configurable on `Application`
This is useful for testing. Not only do the logs polute the log output,
but we were also unable to have more than one test spawning an
`Application`, because each call to `start` was setting the global
subscriber, what can only be done once. Passing a boolean parameter to
`start` we can now disable tracing in the tests, while retaining it in
production.